requirements.txt
farm-haystack[inference]
pandas